requirement.txt
openai
numpy
faiss-gpu